+++ Horologion/src/command_base.hpp
@@ -1,17 +1,6 @@
 #pragma once
 
-#include <ctime>
-#include <string>
-#include <vector>
-
-struct Configs {
-    std::vector<std::pair<std::string, std::string>> commands = {};
-    std::string suspend_type = "mem";
-
-    struct tm time_wake = { 0 };
-    struct tm time_run_cmd = { 0 };
-    struct tm time_sleep = { 0 };
-};
+#include "parse_config_file.hpp"
 
 class CommandBase {
 public:
